begin skin recode
