+++ WeaponMechanics/src/main/java/me/deecaad/weaponmechanics/weapon/skin/RelativeSkin.java
@@ -1,0 +1,9 @@
+package me.deecaad.weaponmechanics.weapon.skin;
+
+import java.util.OptionalInt;
+
+public class RelativeSkin {
+
+    private final OptionalInt addDurability;
+    private final OptionalInt addCustomModelData;
+}
